Added helper for creating api response handlers
